put everything about the rendering api in a viewer/ folder
all the network based stuff and interaction with mineflayer is in lib

this viewer/ folder is self sufficient and has a simple API
