+++ lib/index.js
@@ -3,7 +3,7 @@
 global.THREE = require('three')
 require('three/examples/js/controls/OrbitControls')
 
-const { Viewer } = require('./viewer')
+const { Viewer } = require('../viewer')
 
 const io = require('socket.io-client')
 const socket = io()
