+++ lib/mineflayer.js
@@ -1,7 +1,7 @@
 const { Vec3 } = require('vec3')
 const EventEmitter = require('events')
 const { RaycastIterator } = require('mineflayer/lib/iterators')
-const { WorldView } = require('./worldView')
+const { WorldView } = require('../viewer')
 
 module.exports = (bot, { viewDistance = 6, firstPerson = false, port = 3000 }) => {
   const express = require('express')
