+++ package.json
@@ -7,7 +7,7 @@
     "test": "jest --verbose --runInBand --forceExit",
     "pretest": "npm run lint && require-self",
     "lint": "standard",
-    "prepare": "npm install require-self && require-self && webpack",
+    "prepare": "npm install require-self && require-self && node viewer/prerender.js && cp -R viewer/public/* public/ && webpack",
     "fix": "standard --fix"
   },
   "author": "PrismarineJS",
@@ -35,6 +35,7 @@
     "webpack-cli": "^4.2.0"
   },
   "devDependencies": {
+    "http-server": "^0.12.3",
     "jest": "^26.6.3",
     "jest-puppeteer": "^4.4.0",
     "minecraft-wrap": "^1.3.0",
