+++ viewer/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  Viewer: require('./lib/viewer').Viewer,
+  WorldView: require('./lib/worldView').WorldView,
+  getBufferFromStream: require('./lib/simpleUtils').getBufferFromStream
+}
