+++ viewer/lib/simpleUtils.js
