+++ viewer/lib/utils.electron.js
