+++ viewer/lib/utils.js
