+++ viewer/lib/worker.js
