+++ viewer/lib/worldrenderer.js
